Added HomesControllerTest

+++ SensorService/Controllers/HomesController.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Common.Repository;
 using Domain.Entities;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
@@ -30,6 +31,12 @@ namespace SensorService.Controllers
         [HttpGet("{id}")]
         public async Task<IActionResult> Get(int id)
         {
+            var home = await this.homeRepository.GetHome(id);
+            if (home == null)
+            {
+                return NotFound();
+            }
+
             return Ok(await this.homeRepository.GetHome(id));
         }
 
@@ -38,6 +45,11 @@ namespace SensorService.Controllers
         public async Task<IActionResult> Post([FromBody]Home value)
         {
             var insertedHome = await this.homeRepository.AddHome(value);
+            if (insertedHome == null)
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError);
+            }
+
             return Ok(insertedHome);
         }
 
@@ -45,7 +57,11 @@ namespace SensorService.Controllers
         [HttpPut]
         public async Task<IActionResult> Put([FromBody]Home value)
         {
-            await this.homeRepository.EditHome(value);
+            var success = await this.homeRepository.EditHome(value);
+            if (!success)
+            {
+                return BadRequest();
+            }
             return Ok(value);
         }
 
@@ -54,7 +70,17 @@ namespace SensorService.Controllers
         public async Task<IActionResult> Delete(int id)
         {
             var home = await this.homeRepository.GetHome(id);
-            await this.homeRepository.DeleteHome(id);
+            if (home == null)
+            {
+                return NotFound();
+            }
+
+            var success = await this.homeRepository.DeleteHome(id);
+
+            if (!success)
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError);
+            }
 
             return Ok(home);
         }
@@ -63,6 +89,11 @@ namespace SensorService.Controllers
         public async Task<IActionResult> GetRooms(int id)
         {
             var home = await this.homeRepository.GetHome(id);
+            if (home == null)
+            {
+                return NotFound();
+            }
+
             var rooms = await this.homeRepository.GetRooms(home);
             return Ok(rooms);
         }
@@ -70,20 +101,51 @@ namespace SensorService.Controllers
         [HttpPost("{id}/room")]
         public async Task<IActionResult> AddRoom(int id, [FromBody]Room room)
         {
+            var home = await this.homeRepository.GetHome(id);
+            if (home == null)
+            {
+                return NotFound();
+            }
             var insertedRoom = await this.homeRepository.AddRoom(id, room);
+            if (insertedRoom == null)
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError);
+            }
             return Ok(insertedRoom);
         }
 
         [HttpPost("{id}/weather")]
         public async Task<IActionResult> AddWeather(int id, [FromBody]Weather weather)
         {
+            var home = await this.homeRepository.GetHome(id);
+            if (home == null)
+            {
+                return NotFound();
+            }
             var insertedWeather = await this.homeRepository.AddWeather(id, weather);
+            if (insertedWeather == null)
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError);
+            }
+
             return Ok(insertedWeather);
         }
 
         [HttpGet("{id}/weather")]
         public async Task<IActionResult> GetWeather(int id, [FromQuery]DateTime startDate, [FromQuery]DateTime endDate)
         {
+            var invalidDate = new DateTime(1970, 1, 1, 0, 0, 0, DateTimeKind.Utc);
+            if (!ModelState.IsValid || startDate.Equals(endDate) || startDate.Equals(invalidDate) || endDate.Equals(invalidDate))
+            {
+                return BadRequest();
+            }
+
+            var home = await this.homeRepository.GetHome(id);
+            if (home == null)
+            {
+                return NotFound();
+            }
+
             var weather = await this.homeRepository.GetWeather(id, startDate, endDate);
             return Ok(weather);
         }
